sha1 patch for ppcskel
Steve Reid's SHA-1 C implementation is used here and slighty modified
for the SHA-1 Controller in Hollywood.
Some simple testcases showed that it is a bit faster as it would be done
in software, especially on a huge amount of data.
However, there is room for improvement :)

Further this patch uses the AHBPROT issue recently found, so Broadway is
able to acces the SHA-1 Controller *directly*.