add debugfunction for td->flag
finally we get some response from HC :) however our TDs seems to be
bullshit :/

@commit 8dbf7ef007d1d:
I think the problem here was that tailp and headp of the ED structure
was set to the same location. In this case, the HC thinks all TDs are
done! So HC set CLF to null. (search for this somewhere in the ohci specs...)
In fact, just headp must be set.